Set QOS / Thread priority
Surprisingly, it seems that leaving the qos priority to default does not actually lead to a qos of background like the documentation states. I've found that the QOS gets upgraded to user interactive in some cases. This explicitly sets the QOS on all operations to background so we don't affect scroll performance.