Fix runaway NSData (#205)
* Fix runaway NSData

Previously we were generating all the frames of the GIF and dispatching
to write the data to the disk. This presents a problem if the disk backs
up though and it can end up with theoretically every frame in memory. Instead,
generate a frame and write it one frame at a time. Slower, but prevents
this runaway memory situation.

* Early return on error, thans @rahul-malik & @timonus!