Added custom retry strategy
Added ability to set custom request strategy. We have noticed that if
there is no internet connection, framework tries to retry 3 times with
exponential retry strategy, overall about 90 seconds needed to get
network error. We wanted to change that behavior, but noticed that
there is no clear way to do that.