[PINAnimatedImageManager] Refine file cleanup in +initialize.
Question: is +initialize sufficient to perform this work?  Since the class may not be initialized,
even on the next launch of the app, do we need to more proactively clean up (without impacting app start?).

One possibility would be to listen for app launch and suspend notifications rather than +load.