Custom retry strategy (#386)
* Added custom retry strategy

Added ability to set custom request strategy. We have noticed that if
there is no internet connection, framework tries to retry 3 times with
exponential retry strategy, overall about 90 seconds needed to get
network error. We wanted to change that behavior, but noticed that
there is no clear way to do that.

* fixes

* Fixed documentation and moved retry strategy into constructor

* Fixed example test project configs, added tests

* Updated changelog

* reverted project.pbxproj file

* Fixed documentaion

* Removed unused defines, removed retry strategy from constructor, minor fix.

* fixed unit test for retry strategy, fixed retry strategy count issue