Allow clients to configure URLSession URLCache if they choose. (#477)
* Allow clients to configure URLSession URLCache if they choose.

Set NSURLRequestReloadIgnoringLocalCacheData in the default
session instead of per-request. This lets a client configure
a session with a URLCache if they so choose. Also explicitly
set URLCache in the default session to nil to avoid the small
costs associated with the default ephemeral session memory
cache that is never used.

* Update CHANGELOG.md