Set priority of download data tasks at the right time (#490)
Currently the priority is set after a task has been scheduled and potentially sent off. It means the priority may not be picked up at all. This diff adds a new behavior that sets the priority right after the task is created.

Also expose an API that allows the priority to be set by Texture.