Fix compile errors when using Makefile (#493)
There were random compile errors happening when using xcode 10 to compile via
the Makefile (like doing `make test`). This is related to the new build system
in xcode 10 and is suspected to be happening because of some sort of race
condition during compilation. The error revolved around multiple Targets
creating the same directories for framework dependencies. I have a feeling this
needs to be solved by Carthage since it sets up the dependencies, but as an
immediate solution we can set the build settings to use the legacy build
system.