On iOS, decode WebP images directly into a buffer Core Animation likes
Without this change, the images we output aren't compatible with the GPU and so CA has to copy them all on the main thread at commit time, which means stutters. Plus memory bloat.

The difference here is so substantial that you may seriously want to consider re-running a WebP experiment if you did in the past and bailed on it.